AA add forklift and timeout endpoints

+++ backend/src/main/java/by/alexander/backend/service/impl/ForkliftServiceImpl.java
@@ -3,10 +3,13 @@ package by.alexander.backend.service.impl;
 import by.alexander.backend.auth.CustomUserDetails;
 import by.alexander.backend.auth.JwtUserDetailsService;
 import by.alexander.backend.dao.ForkliftRepository;
+import by.alexander.backend.dao.ForkliftTimeoutRepository;
 import by.alexander.backend.dao.UserRepository;
 import by.alexander.backend.dto.ForkliftDto;
 import by.alexander.backend.entity.Forklift;
 import by.alexander.backend.entity.User;
+import by.alexander.backend.exception.BadRequestException;
+import by.alexander.backend.exception.ResourceNotFoundException;
 import by.alexander.backend.mapper.ForkliftMapper;
 import by.alexander.backend.service.ForkliftService;
 import lombok.AccessLevel;
@@ -16,6 +19,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+import static by.alexander.backend.util.ConstantsUtil.*;
+
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
@@ -24,10 +29,11 @@ public class ForkliftServiceImpl implements ForkliftService {
     ForkliftRepository forkliftRepository;
     ForkliftMapper forkliftMapper;
     UserRepository userRepository;
+    ForkliftTimeoutRepository forkliftTimeoutRepository;
 
     public ForkliftDto getById(Long id) {
         Forklift forklift = forkliftRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException(FORKLIFT, ID, id));
         return forkliftMapper.toDto(forklift);
     }
 
@@ -45,7 +51,7 @@ public class ForkliftServiceImpl implements ForkliftService {
 
     public void edit(Long id, ForkliftDto forkliftDto) {
         Forklift forklift = forkliftRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException(FORKLIFT, ID, id));
         User user = getUserById();
         forklift = forkliftMapper.toEntityUpdate(forklift, forkliftDto, user);
         forkliftRepository.save(forklift);
@@ -53,13 +59,17 @@ public class ForkliftServiceImpl implements ForkliftService {
 
     public void delete(Long id) {
         Forklift forklift = forkliftRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException(FORKLIFT, ID, id));
+        boolean isHasTimeout = forkliftTimeoutRepository.existsByForkliftIdAndSolutionDateIsNotNull(id);
+        if (isHasTimeout) {
+            throw new BadRequestException("Forklift has been active timeout");
+        }
         forkliftRepository.deleteById(forklift.getId());
     }
 
     private User getUserById() {
         CustomUserDetails auth = JwtUserDetailsService.getAuthentication();
         return userRepository.findById(auth.getId())
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException(USERS, ID, auth.getId()));
     }
 }
